r600g: wip bank_swizzle scheduler
this probably will never be finished. Turned out: α) r600g have for some
reason 2 schedulers doing the same thing, and this one simply makes a
validation before some time later the 2nd going to run. β) time
measurements of the part that in this scheduler works (it is vector part
scheduling, minus some weirdness of documentation) showed no difference
in μsec granularity for the whole start of Xonotic + loading a level.

NB: the commit been amended many times, and because it is "wip", I didn't
care to save all changes.

Signed-off-by: Constantine Kharlamov <Hi-Angel@yandex.ru>